// src/api/models/WebhookRepresentation.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type WebhookRepresentation = {
    attributes?: Record<string, any>;
    id?: string;
    enabled?: boolean;
    url?: string;
    secret?: string;
    createdBy?: string;
    createdAt?: string;
    realm?: string;
    eventTypes?: Array<string>;
};